[Refactor] #9 - Generation 엔티티 인터페이스 구현 방식으로 변경
