+++ module-api/src/main/kotlin/org/iass/generation/facade/GenerationFacadeImpl.kt
@@ -1,0 +1,20 @@
+package org.iass.generation.facade
+
+import org.iass.generation.dto.request.GenerationCreateRequest
+import org.iass.generation.service.GenerationCommandService
+import org.iass.generation.service.GenerationQueryService
+import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
+import java.net.URI
+
+@Service
+@Transactional(readOnly = true)
+class GenerationFacadeImpl(
+	private val generationCommandService: GenerationCommandService,
+	private val generationQueryService: GenerationQueryService
+) : GenerationFacade {
+	override fun create(request: GenerationCreateRequest): URI {
+		val generation = generationCommandService.create(request)
+		return URI.create("/api/generation/${generation.id}")
+	}
+}
