+++ module-api/src/main/kotlin/org/iass/generation/service/GenerationQueryService.kt
@@ -1,8 +1,3 @@
 package org.iass.generation.service
 
-import org.springframework.stereotype.Service
-import org.springframework.transaction.annotation.Transactional
-
-@Service
-@Transactional(readOnly = true)
-class GenerationQueryService
+interface GenerationQueryService
