+++ module-api/src/main/kotlin/org/iass/generation/service/GenerationQueryServiceImpl.kt
@@ -1,0 +1,8 @@
+package org.iass.generation.service
+
+import org.springframework.stereotype.Service
+import org.springframework.transaction.annotation.Transactional
+
+@Service
+@Transactional(readOnly = true)
+class GenerationQueryServiceImpl() : GenerationQueryService
